layout
changed spacing on the top left input area. Was a bit too spacious above the input, due to the inflated input text area. and below, had the same padding as above, though it is home to a smaller font size and pictogram. so reduced the space to make it feel more appropriate for the smaller stuff